feat(admin): new dashboard detail/edit page admin UI [EP-2196]
